Fix Web Workers example page title
The Deno By Example page for Web Workers has the title of "Web workers".

This commit corrects this to correct spelling "Web Workers", with a capital W in each word, so that it will appear with the correct spelling on the title of this page and on the link on the Deno By Example page. This commit also corrects the same spelling once in the body of the page, and corrects the spelling of "javascript" to "JavaScript".