Fix Deno By Example page titles
This commit corrects the titles of the following Deno By Example pages so they it will appear with the correct spelling on the title of this page and on the link on the Deno By Example page:

"Web assembly" changed to "WebAssembly".
"Web workers" changed to "Web Workers".

This commit also changes one occurrence of "javascript" to "JavaScript".